+++ src/com/djj/UAV/Tsp_interface.java
@@ -3,5 +3,5 @@ package com.djj.UAV;
 import java.util.List;
 
 public interface Tsp_interface {
-	public void find_best(int [][]Graph,List<List<Integer>> list,int num_target,int num_Uav);
+	public void find_best(double [][]Graph,List<List<Integer>> list,int num_target,int num_Uav);
 }
